requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
faster-whisper>=1.0.0
gTTS>=2.5.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
fakeredis>=2.21.0
respx>=0.21.0
aiofiles>=24.1.0
# Agent framework (optional — works without it via simple routing)
langgraph>=0.2.0
# Vector search (optional — RAG unavailable without it)
chromadb>=0.5.0
# PDF generation (optional — document generation unavailable without it)
fpdf2>=2.7.0
# JWT authentication (optional — token auth unavailable without it)
PyJWT>=2.8.0
